+++ tests/base_tests/test_function_user.py
@@ -13,14 +13,27 @@ pytestmark = [pytest.mark.asyncio]
 app = Application(cookies_secret=b'123456', permission=None)
 
 
+class SimpleUser(dict, BaseUser):
+    pass
+
+
+class UserViewMixin(BaseUserViewMixin):
+    async def get_current_user(self):
+        return SimpleUser({'name': 'qiuye'})
+
+
 @app.route.get('base')
-def for_test(request: RequestView):
-    request.finish(RETCODE.SUCCESS, 111)
+def for_test(req: RequestView):
+    req._.user = req.current_user
+    req.finish(RETCODE.SUCCESS)
 
 
 app.prepare()
+app.set_user_mixin_class(UserViewMixin)
 
 
-async def test_func_base():
+async def test_func_with_user():
     view = await invoke_interface(app, for_test)
     assert view.retcode == RETCODE.SUCCESS
+    assert view._.user
+    assert view._.user['name'] == 'qiuye'